+++ src/app/admin/artist_booking/all/(components)/ArtistBookingTableListing.tsx
@@ -5,8 +5,8 @@ import { PaginatedResponseType } from '@/core/types/responseTypes';
 import AlertDialog from '@/core/ui/components/AlertDialog';
 import PaginationNav from '@/core/ui/components/Pagination';
 import { Button, TableCard, tableStyles } from '@/core/ui/zenbuddha/src';
-import artistBookingApi from '@/modules/artist_booking/artist_bookingApi';
-import { ArtitstBookingType } from '@/modules/artist_booking/artist_bookingType';
+import artistBookingApi from '@/modules/artists/artist_bookingApi';
+import { ArtitstBookingType } from '@/modules/artists/artistsType';
 import { format, parseISO } from 'date-fns';
 import { toZonedTime } from 'date-fns-tz';
 import { Clock, Eye, PencilSimpleLine, TrashSimple } from 'phosphor-react';
