+++ src/modules/artists/artistsType.ts
@@ -1,21 +1,7 @@
+import { selectorDataSchema } from '@/core/types/selectorType';
 import { nonempty } from '@/core/utils/formUtlis';
+import { documentFile, imageFile } from '@/core/utils/helper';
 import { z } from 'zod';
-
-const imageFile = z.instanceof(File).refine(
-  (file) => {
-    const acceptedImageTypes = [
-      'image/jpeg',
-      'image/png',
-      'image/gif',
-      'image/bmp',
-      'image/webp',
-    ];
-    return acceptedImageTypes.includes(file.type);
-  },
-  {
-    message: 'Invalid file type. Only image files are allowed.',
-  }
-);
 
 export const artistsSchema = z.object({
   id: z.number().optional().nullable(),
@@ -24,8 +10,34 @@ export const artistsSchema = z.object({
   bio: z.string(),
   profile_picture: imageFile.optional().nullable(),
 });
+export const artitstBookingSchema = z.object({
+  id: z.number().optional().nullable(),
+  full_name: z.string().pipe(nonempty),
+  phone: z.string().optional(),
+  event_date: z.date().optional().nullable(),
+  event_time: z.string().optional().nullable(),
+  event_type: z.string().optional(),
+  email: z.string().email(),
+  location: z.string().optional(),
+  document: documentFile.optional().nullable(),
+  info: z.string().optional(),
+  artist: selectorDataSchema,
+});
 
+export type ArtitstBookingSchemaType = z.infer<typeof artitstBookingSchema>;
 export type ArtistsSchemaType = z.infer<typeof artistsSchema>;
+export type ArtitstBookingType = {
+  id: number;
+  full_name: string;
+  event_date?: string;
+  event_type?: string;
+  email?: string;
+  phone?: string;
+  location?: string;
+  document?: string;
+  info?: string;
+  artist?: ArtistsType;
+};
 
 export type ArtistsType = {
   id: number;
